requirements.txt
Flask
requests
beautifulsoup4
markitdown[pdf]
python-dotenv
gunicorn
tabula-py
pandas
python-docx
supabase